refactor: api files refactor and removing unused files
